New Skinning test with simple unweighted bones and includes in shaders.
The Shader class now support a #include preprocessing instruction that
allows to build the text file of a shader from several sources. An
include path is also provided to setup a list of usual paths for
includes.

The Skinning test and the Bone class do work. Added some shaders that
uses them.