Added a DynPointMesh.
The DynPointMesh allow to setup a set of points and to update
dynamically their position and colors. Its the first dynamic mesh of
SOFA, but other may come (and refinements of the API, probably).

Also added a brownianMotion() method to the NumberSeq.

Also allow to set the usage hint of ArrayBuffers (static or dynamic
draw), and also from the VertexArray.