Abandonned the Nio versions of Vectors and Points after some profiling.
This was a bad idea for a lot of reasons. The new classes in NumberSeq,
Point, and Vector are now quite simpler and a lot faster (data is no
more a method, but an array !!). It is far less costly to convert the
arrays back to Nio only when needed and do all the calculus on arrays.

- also lots of micro optimisations in VES, applied in the Vector and
Point classes that allow very great gains in speed.