Start of work on render layers for avatars.
- Avatar now have hasLayer and layer attributes that allow optional
off-screen rendering.
- AvatarRender now provides pushLayer(), popLayer() and disposeLayer()
that allows to divert the render toward an automatically allocated
off-screen buffer that can later be rendered as a texture.
- Times are divided by 4 but the rendering cannot use FSAA, therefore
this technique is limited to high-dpi screens (or we have to do ourself
the FSAA by doubling the size of the off-screen buffers ?).
- Space now supports a viewport stack as for projection and model view
matrices. Handy for the off-screen buffers.
- UIList has some tests for off-screen rendering of items.