New type face resource.
- Added a GLTypeFace class that groups font sizes of a type face.
- Finalized FontResource, but changed it into TypeFaceResource in the
Library.
- The Screen now provides an utility font() method allowing to get a
font of the correct size in millimeters even if the surface changes
physical screen.
- The Screen now contains a « pixelSpace » in addition to the space,
where each integer position maps to a real pixel of the output.
- This pixel space is updated in AvatarRender when pushing or popping
layers.
- This pixel space also allows to render text directly during rendering
(instead of using a secondary pass), necessary when rendering into a
layer (slower than rendering after in one pass, but the layer gains
will make this invisible).