Reworked completely the Mesh trait.
- The way Mesh works changed radically. Before Mesh was a factory for
VertexArrays. Now Mesh and its descendants are classes used directly to
draw geometry.
- The mesh class encapsulate vertex attribute data and optional indices.
- It provides draw commands.
- The way the data is stored depends on the implementation, it can work
by copying data to the GL or mapping GL memory or streaming it.
- The Mesh trait now defines a begin/end mechanism to handle changes on
the geometry.
- The Mesh trait now hides the vertex array and the shader bound to it.
But it remember them.
- Mesh also now stores the elements directly. Sub-classes are far
simpler.
- Mesh now allows sub-classes to choose when to add vertex attributes
depending on needs.
- Mesh now stores the vertex attributes using MeshAttribute, an
abstract class allowing several implementations to store the data, send
it to the GL, etc. Idem for MeshElement.
- Meshes can share their vertex attributes.

- ArrayBuffer and ElementBuffer now allow construction of buffers
without uploading data first.

- ColladaGeometry now exports a TrianglesMesh, EditableMesh will
disappear.